+++ Editor/DG_Transfer_Colliders.cs
@@ -5,59 +5,73 @@ namespace DG_Toolchain
 {
     public class DG_Transfer_Colliders : EditorWindow
     {
-        static private string childName = "Empty";
-
+        
         static public void OnGUI()
         {
-            EditorGUILayout.HelpBox("If there is no child with the provided name new one will be created instead.", MessageType.Info);
-            childName = EditorGUILayout.TextField("Child Name:", childName);
-            EditorGUI.BeginDisabledGroup(Selection.activeGameObject == null || childName == string.Empty);
-            if (GUILayout.Button("Transfer Colliders To Child"))
+            GameObject selected = Selection.activeGameObject;
+            GUILayout.BeginHorizontal();
+            EditorGUI.BeginDisabledGroup((selected == null) || (selected.transform.parent == null) || (selected.transform.parent.GetComponent<Collider>() == null));
+            if (GUILayout.Button("To Selected Child", UnityEditor.EditorStyles.miniButtonLeft)) 
             {
-                TransferColliders();
+                TransferCollidersToChild();
             }
             EditorGUI.EndDisabledGroup();
+
+            EditorGUI.BeginDisabledGroup((selected == null) || !HasComponentInChildren<Collider>(selected));
+            if (GUILayout.Button("To Selected Parent", UnityEditor.EditorStyles.miniButtonRight))
+            {
+                TransferCollidersToParent();
+            }
+            EditorGUI.EndDisabledGroup();
+            GUILayout.EndHorizontal();     
         }
 
-        static private void TransferColliders()
+        static bool HasComponentInChildren<T>(GameObject obj) where T : Component
+        {
+            bool result = false;
+            foreach (Transform childTransform in obj.transform)
+            {
+                if (childTransform.GetComponent<T>() != null)
+                {
+                    result = true;
+                    break;
+                }
+            }
+            return result;
+        }
+
+        static private void TransferCollidersToParent()
         {
 
-            Transform parentTransform = Selection.activeGameObject.transform;
-            Transform childTransform = FindOrCreateChild(parentTransform, childName);
+            Collider[] childColliders = Selection.activeGameObject.GetComponentsInChildren<Collider>(true);
 
-            if (childTransform != null)
+            foreach (Collider childCollider in childColliders)
             {
+                Undo.RecordObject(childCollider, "Transfer Collider to Parent");
+                Collider newCollider = Undo.AddComponent(Selection.activeGameObject.gameObject, childCollider.GetType()) as Collider;
+                EditorUtility.CopySerialized(childCollider, newCollider);
+                Undo.DestroyObjectImmediate(childCollider);
+            }
+        }
+
+        static private void TransferCollidersToChild()
+        {
+
+            Transform childTransform = Selection.activeGameObject.transform;
+            Transform parentTransform = Selection.activeGameObject.transform.parent;
+
                 Collider[] colliders = parentTransform.GetComponents<Collider>();
 
                 foreach (Collider collider in colliders)
                 {
                     if (collider != null)
                     {
-                        Undo.RecordObject(collider, "Transfer Collider");
+                        Undo.RecordObject(collider, "Transfer Collider to Child");
                         Collider newCollider = Undo.AddComponent(childTransform.gameObject, collider.GetType()) as Collider;
                         EditorUtility.CopySerialized(collider, newCollider);
                         Undo.DestroyObjectImmediate(collider);
                     }
                 }
-            }
-        }
-
-        static private Transform FindOrCreateChild(Transform parent, string childName)
-        {
-            Transform child = parent.Find(childName);
-
-            if (child == null)
-            {
-                GameObject newChild = new GameObject(childName);
-                newChild.transform.SetParent(parent);
-                newChild.transform.localPosition = Vector3.zero;
-                newChild.transform.localRotation = Quaternion.identity;
-                newChild.transform.localScale = Vector3.one;
-
-                return newChild.transform;
-            }
-
-            return child;
         }
     }
 }