+++ Changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2023-12-13
+
+### Added
+- Collider manipulation scripts.
+- LOD group manipulation scripts.
+
+### Removed
+- `Generate prefabs` script. The standard method of generating prefabs by dragging multiple gameobjects into the Project window seems to be more efficient. 
+
 ## [1.0.1] - 2023-12-05
 
 ### Added